pull out helper function from validation
